Refactors fees estimation code as a separate interoperable function
