+++ src/component/upload_show.jsx
@@ -2,11 +2,10 @@ import { React, useState, useRef } from 'react';
 import ArDB from 'ardb';
 import { CONTRACT_SRC, FEE_MULTIPLIER, arweave, languages_en, languages_zh, categories_en, categories_zh, smartweave } from '../utils/arweave.js'
 import { genetateFactoryState } from '../utils/initStateGen.js';
-import { swal, fetchWalletAddress } from '../utils/shorthands.js';
+import { processFile, fetchWalletAddress, userHasEnoughAR } from '../utils/shorthands.js';
 
 import Swal from 'sweetalert2';
 import { useTranslation } from 'react-i18next';
-import { getStorageTable } from 'arweave-fees.js';
 
 const ardb = new ArDB(arweave)
 
@@ -41,27 +40,7 @@ export default function UploadShow() {
     return tx.id
   }
 
-  function readFileAsync(file) {
-    return new Promise((resolve, reject) => {
-      let reader = new FileReader();
-      reader.onload = () => {
-        resolve(reader.result);
-      };
-      reader.onerror = reject;
-      reader.readAsArrayBuffer(file);
-    })
-  }
-
-  async function processFile(file) {
-    try {
-      let contentBuffer = await readFileAsync(file);
-
-      return contentBuffer
-    } catch (err) {
-      console.log(err);
-    }
-  }
-
+  
   const uploadShow = async (show) => {
     Swal.fire({
       title: t("uploadshow.swal.uploading.title"),
@@ -205,21 +184,12 @@ export default function UploadShow() {
     showObj.lang = podcastLanguage
     // upload cover, send all to Arweave
     let cover = await processFile(podcastCover)
-    const storagePrices = await getStorageTable()
-    let address = await fetchWalletAddress()
-    let failText = 'uploadshow.swal.uploadfailed.'
-    if (!address) return swal(t, 'error', failText + 'cantfindaddress')
+    let showObjSize = JSON.stringify(showObj).length
+    let bytes = cover.byteLength + showObjSize + coverFileType.length
 
-    let balance = await arweave.wallets.getBalance(address).then((balance) => balance)
-    let avgStoragePriceKB = storagePrices['KB'] ? storagePrices['KB'].ar : 0
-
-    if (avgStoragePriceKB == 0) return swal(t, 'error', failText + 'cantfetchprices')
-
-    let showObjSize = JSON.stringify(showObj).length // in bytes
-    let cost = (Math.max(cover.byteLength + showObjSize, 1) / 1024) * avgStoragePriceKB
-
-    if (balance >= cost) await uploadToArweave(cover, coverFileType, showObj)
-    else return swal(t, 'error', failText + 'lowbalance', `${cost}`.split("", 6).join('') + ' AR')
+    if (userHasEnoughAR(t, bytes) === "all good") {
+      await uploadToArweave(cover, coverFileType, showObj)
+    }
   }
 
   const languageOptions = () => {
